requirements.txt
pandas==2.1.4
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
flask==3.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
